add view logs, fix auth and editUser

+++ src/sample/Settings.java
@@ -3,6 +3,6 @@ package sample;
 public class Settings {
 
     public static int Tried_Pass = 3;
-    public static int TimeBlock = 5; //мин 15
-    public static int TimeOut = 5; //мин
+    public static long TimeBlock = 5; //мин 15
+    public static long TimeOut = 5; //мин
 }
